Add flash_append method to session_lib
This lets you append to an array in the flash data, creating one first
if it does not exist. This is particularly useful in cases such as
errors, $this->session_lib->flash_append('errors', 'foo bar'); lets you
append errors from anywhere in the code without worrying about whether
the 'errors' array exists in the flash data.